docs/requirements.txt
autoconf==0.7.2
autofit==0.70.0
autoarray==0.15.0
autogalaxy==0.16.0
astropy>=3.0.0
nbsphinx>=0.7.0
nbsphinx-link>=1.3.0
numpydoc>=1.0.0
matplotlib>=3.0.3
numpy>=1.17.2
ipython>=7.0.1
ipykernel>=5.1.0
jupyter_client>=5.2.3
pygments==2.4.1
pyprojroot>=0.2.0